apply ruff format

+++ tests/test_asdf_in_fits.py
@@ -13,21 +13,21 @@ def test_write_linked(tmp_path):
     hdulist = fits.HDUList()
     sci = np.arange(512, dtype=float)
     dq = np.arange(512, dtype=float) + 1
-    hdulist.append(fits.ImageHDU(sci, name='SCI'))
-    hdulist.append(fits.ImageHDU(dq, name='DQ'))
+    hdulist.append(fits.ImageHDU(sci, name="SCI"))
+    hdulist.append(fits.ImageHDU(dq, name="DQ"))
 
     tree = {
-        'meta': {
-            'foo': 'bar',
+        "meta": {
+            "foo": "bar",
         },
-        'model': {
-            'sci': {
-                'data': hdulist['SCI'].data,
+        "model": {
+            "sci": {
+                "data": hdulist["SCI"].data,
             },
-            'dq': {
-                'data': hdulist['DQ'].data,
-            }
-        }
+            "dq": {
+                "data": hdulist["DQ"].data,
+            },
+        },
     }
 
     fn = tmp_path / "test.fits"
@@ -36,34 +36,34 @@ def test_write_linked(tmp_path):
 
     # confirm it matches
     with asdf_in_fits.open(fn) as af:
-        assert_array_almost_equal(af['model']['sci']['data'], sci)
-        assert_array_almost_equal(af['model']['dq']['data'], dq)
-        assert af['meta']['foo'] == 'bar'
+        assert_array_almost_equal(af["model"]["sci"]["data"], sci)
+        assert_array_almost_equal(af["model"]["dq"]["data"], dq)
+        assert af["meta"]["foo"] == "bar"
 
     with fits.open(fn) as read_hdulist:
         # hdu should have sci, dq, and ASDF
         assert len(read_hdulist) == 3
         # check asdf extension has no blocks by looking for bytes
         # after the end of the yaml document for the tree
-        bs = read_hdulist['ASDF'].data['ASDF_METADATA'].tobytes()
-        assert len(bs.strip().split(b'...')[1]) == 0
+        bs = read_hdulist["ASDF"].data["ASDF_METADATA"].tobytes()
+        assert len(bs.strip().split(b"...")[1]) == 0
 
 
 def test_write_asdf_in_fits_no_hdulist(tmp_path):
     sci = np.arange(512, dtype=float)
     dq = np.arange(512, dtype=float) + 1
     tree = {
-        'meta': {
-            'foo': 'bar',
+        "meta": {
+            "foo": "bar",
         },
-        'model': {
-            'sci': {
-                'data': sci,
+        "model": {
+            "sci": {
+                "data": sci,
             },
-            'dq': {
-                'data': dq,
-            }
-        }
+            "dq": {
+                "data": dq,
+            },
+        },
     }
     fn = tmp_path / "test.fits"
 
@@ -71,16 +71,16 @@ def test_write_asdf_in_fits_no_hdulist(tmp_path):
 
     # confirm it matches
     with asdf_in_fits.open(fn) as af:
-        assert_array_almost_equal(af['model']['sci']['data'], sci)
-        assert_array_almost_equal(af['model']['dq']['data'], dq)
-        assert af['meta']['foo'] == 'bar'
+        assert_array_almost_equal(af["model"]["sci"]["data"], sci)
+        assert_array_almost_equal(af["model"]["dq"]["data"], dq)
+        assert af["meta"]["foo"] == "bar"
 
     with fits.open(fn) as read_hdulist:
         # hdu should have primary and ASDF
         assert len(read_hdulist) == 2
         # check asdf extension has blocks by looking at the block index
-        bs = read_hdulist['ASDF'].data['ASDF_METADATA'].tobytes()
-        block_index_bs = bs.split(b'BLOCK INDEX')[1].strip()
+        bs = read_hdulist["ASDF"].data["ASDF_METADATA"].tobytes()
+        block_index_bs = bs.split(b"BLOCK INDEX")[1].strip()
         block_offsets = yaml.load(block_index_bs, yaml.SafeLoader)
         assert len(block_offsets) == 2
 
@@ -93,17 +93,17 @@ def test_write_asdf_in_fits_partial_hdulist(tmp_path):
     sci = np.arange(512, dtype=float)
     dq = np.arange(512, dtype=float) + 1
     tree = {
-        'meta': {
-            'foo': 'bar',
+        "meta": {
+            "foo": "bar",
         },
-        'model': {
-            'sci': {
-                'data': sci,
+        "model": {
+            "sci": {
+                "data": sci,
             },
-            'dq': {
-                'data': dq,
-            }
-        }
+            "dq": {
+                "data": dq,
+            },
+        },
     }
     fn = tmp_path / "test.fits"
 
@@ -111,23 +111,23 @@ def test_write_asdf_in_fits_partial_hdulist(tmp_path):
 
     # confirm it matches
     with asdf_in_fits.open(fn) as af:
-        assert_array_almost_equal(af['model']['sci']['data'], sci)
-        assert_array_almost_equal(af['model']['dq']['data'], dq)
-        assert af['meta']['foo'] == 'bar'
+        assert_array_almost_equal(af["model"]["sci"]["data"], sci)
+        assert_array_almost_equal(af["model"]["dq"]["data"], dq)
+        assert af["meta"]["foo"] == "bar"
 
     with fits.open(fn) as read_hdulist:
         # hdu should have primary and ASDF
         assert len(read_hdulist) == 2
         # check asdf extension has blocks by looking at the block index
-        bs = read_hdulist['ASDF'].data['ASDF_METADATA'].tobytes()
-        block_index_bs = bs.split(b'BLOCK INDEX')[1].strip()
+        bs = read_hdulist["ASDF"].data["ASDF_METADATA"].tobytes()
+        block_index_bs = bs.split(b"BLOCK INDEX")[1].strip()
         block_offsets = yaml.load(block_index_bs, yaml.SafeLoader)
         assert len(block_offsets) == 2
 
 
 @pytest.fixture
 def test_array():
-    return np.arange(1000, dtype='f4').reshape((100, 10))
+    return np.arange(1000, dtype="f4").reshape((100, 10))
 
 
 @pytest.fixture
@@ -143,7 +143,7 @@ def test_open_asdf_in_fits(saved_array_model, test_array):
     # read it back (not in a context)
     af = asdf_in_fits.open(saved_array_model)
     # confirm it matches
-    assert_array_almost_equal(af['data'], test_array)
+    assert_array_almost_equal(af["data"], test_array)
     af.close()
 
 
@@ -151,7 +151,7 @@ def test_open_asdf_in_fits_context(saved_array_model, test_array):
     # read it back as a context
     with asdf_in_fits.open(saved_array_model) as af:
         # confirm it matches
-        assert_array_almost_equal(af['data'], test_array)
+        assert_array_almost_equal(af["data"], test_array)
 
 
 def test_open_asdf_in_fits_hdu(saved_array_model, test_array):
@@ -162,6 +162,6 @@ def test_open_asdf_in_fits_hdu(saved_array_model, test_array):
     """
     with fits.open(saved_array_model) as hdu:
         with asdf_in_fits.open(hdu) as af:
-            assert_array_almost_equal(af['data'], test_array)
+            assert_array_almost_equal(af["data"], test_array)
         # make sure file was not closed with context
-        assert not hdu.fileinfo(0)['file'].closed
+        assert not hdu.fileinfo(0)["file"].closed
